// src/api/generated/model/slackMessageResponseDto.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Dev Blackbox API
 * OpenAPI spec version: 0.1.1
 */

export interface SlackMessageResponseDto {
  id: number;
  target_date: string;
  channel_id: string;
  channel_name: string;
  message_ts: string;
  message_text: string;
  thread_ts: string | null;
}