// src/api/generated/model/githubPullRequestEventPayload.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Dev Blackbox API
 * OpenAPI spec version: 0.1.1
 */
import type { GithubPullRequestModel } from './githubPullRequestModel';

export interface GithubPullRequestEventPayload {
  action: string;
  number: number;
  pull_request: GithubPullRequestModel;
}